+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Reply to see content Changelog
 
+### 1.0.2 - 2020-04-17
+- Added permission to see all content for groups. 
+- Changed code for review first post.
+- Changed search code for first post hide when using permission.
+
+### 1.0.1 - 2020-04-16
+- Fix code for viewtopic.
+- Added permission to see first post content. 
+
 ### 1.0.0 - 2020-04-15
 
 - First release.